pyproject.toml
[tool.poetry]
name = "linktree"
version = "0.1.0"
description = ""
authors = ["gonzalogacc <gonzalogacc@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.97.0"
aiohttp = "^3.8.4"
python-jose = "^3.3.0"
uvicorn = "^0.22.0"
watchgod = "^0.8.2"
watchfiles = "^0.19.0"
boto3 = "^1.26.155"
pynamodb = "^5.5.0"
requests = "^2.31.0"
pytest = "^7.3.2"
moto = "^4.1.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
